Make the temp subroutine generic.
Wrap the temp subroutine in botsma.pl around Botsma::Common::temp, to
give a different output for users within the %users hash. Before this
hash was implemented, there were different functions for different
users.

The 'different output' consists of the difference between a preferred
weather station (as stored in the hash) and the weather station of
Twenthe.