Implement crude user preference system.
Added a simple system to store the preferences of a user. The
preferences are persistent with reloads of the bot, using the core
module Storable.

Also made some slight adjustments to the 'regen' and 'temp' wrappers.

For 'temp' this means that storing a preference 'Twenthe' will not
compare the weather station to itself.

For 'regen' it means that a point of interest, stored as a preference,
will also work now.