Implement nearest weather station search.
Automatically search the nearest weather station when using the 'temp'
function, based on the 'location' preference. This works by looking up
the coordinates of the location, getting x and y coordinates from a
Mercator projection, and finding the nearest neighbor of this point by
making use of a kd-tree in which the weather stations are stored.

Unfortunately doesn't make use of an R-tree as suggested by @akaIDIOT,
but I think it's still sufficient :-)

A user system was already in place, but the 'wstation' preference is now
ignored. We will delete it in the future.

Closes #1.