Possibly use nicknames for 'regen' and 'temp'.
Because 'place' is also used to check whether a location is valid when
saving it to a preference setting, I couldn't add a generic 'nick location
checker' in this function.

Also, because 'temp' expects a weather station and not GPS coordinates
like 'regen' does, handling a nickname is different for both.

So, the solution is a bit dirtier than seemed to be possible at first,
but it does work...

Closes #3.